refactor(github): Extract `utils/github.ts` from datasources
