+++ lib/modules/datasource/github-releases/cache/cache-base.ts
@@ -3,13 +3,13 @@ import { DateTime, DurationLikeObject } from 'luxon';
 import { logger } from '../../../../logger';
 import * as memCache from '../../../../util/cache/memory';
 import * as packageCache from '../../../../util/cache/package';
+import { getApiBaseUrl } from '../../../../util/github';
 import type {
   GithubGraphqlResponse,
   GithubHttp,
   GithubHttpOptions,
 } from '../../../../util/http/github';
 import type { GetReleasesConfig } from '../../types';
-import { getApiBaseUrl } from '../common';
 import type {
   CacheOptions,
   ChangelogRelease,
