+++ lib/modules/datasource/go/common.ts
@@ -1,5 +1,5 @@
+import { getSourceUrl as githubSourceUrl } from '../../../util/github';
 import { BitBucketTagsDatasource } from '../bitbucket-tags';
-import { getSourceUrl as githubSourceUrl } from '../github-releases/common';
 import { GithubTagsDatasource } from '../github-tags';
 import { GitlabTagsDatasource } from '../gitlab-tags';
 import { getSourceUrl as gitlabSourceUrl } from '../gitlab-tags/util';
